Revert "unfuck"
This reverts commit 2e9425af445872046d4da738ea3f4212a638500a.

+++ osu.Game.Rulesets.Vitaru/Ruleset/Objects/HitObjects/Drawables/DrawableLaser.cs
@@ -74,7 +74,7 @@ namespace osu.Game.Rulesets.Vitaru.Ruleset.Objects.HitObjects.Drawables
 
             Origin = Anchor.BottomCentre;
             Position = HitObject.Position;
-            Rotation =(float)HitObject.Angle;
+            Rotation =HitObject.Angle;
 
             Alpha = 1;
